feat(runtimes): extract runkon-runtimes crate from conductor-core (RFC 007)
Extract the agent runtime layer into a new workspace crate
that depends only on serde, serde_json, thiserror, chrono, ulid, libc,
tempfile, and tracing. This decouples the runtime harness from conductor's
full domain (database, config, workflow engine).

Key changes:
- New crate: runkon-runtimes with AgentRuntime trait + claude/cli/script impls
- RuntimeRequest now carries tracker/event_sink/log_path instead of db_path
- Added RunTracker + RunEventSink traits for lifecycle DB decoupling
- Added SqliteHostAdapter in conductor-core to bridge traits over AgentManager
- conductor-core re-exports runkon-runtimes types and provides backward-compat
  wrappers for spawn_headless, try_spawn_headless_run, drain_stream_json
- Updated all call sites in conductor-core, conductor-tui, conductor-web
- Updated integration tests for CliRuntime and ScriptRuntime

All workspace tests pass.
