+++ conductor-core/src/agent/status.rs
@@ -1,5 +1,8 @@
 use serde::{Deserialize, Serialize};
 
+// Re-export moved types from runkon-runtimes
+pub use runkon_runtimes::{AgentRunStatus, StepStatus};
+
 /// Default error message used when the agent reports an error without a message.
 pub const DEFAULT_AGENT_ERROR_MSG: &str = "Claude reported an error";
 
@@ -78,47 +81,6 @@ pub(crate) fn truncate_utf8(s: &str, max_bytes: usize) -> &str {
     &s[..end]
 }
 
-/// Status of an agent run.
-#[cfg_attr(feature = "openapi", derive(utoipa::ToSchema))]
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum AgentRunStatus {
-    Running,
-    WaitingForFeedback,
-    Completed,
-    Failed,
-    Cancelled,
-}
-
-impl std::fmt::Display for AgentRunStatus {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let s = match self {
-            Self::Running => "running",
-            Self::WaitingForFeedback => "waiting_for_feedback",
-            Self::Completed => "completed",
-            Self::Failed => "failed",
-            Self::Cancelled => "cancelled",
-        };
-        write!(f, "{s}")
-    }
-}
-
-impl std::str::FromStr for AgentRunStatus {
-    type Err = String;
-    fn from_str(s: &str) -> std::result::Result<Self, Self::Err> {
-        match s {
-            "running" => Ok(Self::Running),
-            "waiting_for_feedback" => Ok(Self::WaitingForFeedback),
-            "completed" => Ok(Self::Completed),
-            "failed" => Ok(Self::Failed),
-            "cancelled" => Ok(Self::Cancelled),
-            _ => Err(format!("unknown AgentRunStatus: {s}")),
-        }
-    }
-}
-
-crate::impl_sql_enum!(AgentRunStatus);
-
 /// Status of a human-in-the-loop feedback request.
 #[cfg_attr(feature = "openapi", derive(utoipa::ToSchema))]
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
@@ -196,45 +158,6 @@ impl std::str::FromStr for FeedbackType {
 }
 
 crate::impl_sql_enum!(FeedbackType);
-
-/// Status of a single plan step.
-#[cfg_attr(feature = "openapi", derive(utoipa::ToSchema))]
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Default)]
-#[serde(rename_all = "snake_case")]
-pub enum StepStatus {
-    #[default]
-    Pending,
-    InProgress,
-    Completed,
-    Failed,
-}
-
-impl std::fmt::Display for StepStatus {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let s = match self {
-            Self::Pending => "pending",
-            Self::InProgress => "in_progress",
-            Self::Completed => "completed",
-            Self::Failed => "failed",
-        };
-        write!(f, "{s}")
-    }
-}
-
-impl std::str::FromStr for StepStatus {
-    type Err = String;
-    fn from_str(s: &str) -> std::result::Result<Self, Self::Err> {
-        match s {
-            "pending" => Ok(Self::Pending),
-            "in_progress" => Ok(Self::InProgress),
-            "completed" => Ok(Self::Completed),
-            "failed" => Ok(Self::Failed),
-            _ => Err(format!("unknown StepStatus: {s}")),
-        }
-    }
-}
-
-crate::impl_sql_enum!(StepStatus);
 
 #[cfg(test)]
 mod tests {
