+++ conductor-core/src/agent_config.rs
@@ -19,6 +19,9 @@ use serde::Deserialize;
 use crate::error::{ConductorError, Result};
 use crate::text_util::parse_frontmatter;
 
+// Re-export moved types from runkon-runtimes
+pub use runkon_runtimes::{AgentDef, AgentRole};
+
 /// How to locate an agent — either a short name resolved via search order, or
 /// an explicit path relative to the repository root.
 ///
@@ -61,53 +64,6 @@ pub fn default_role() -> String {
 
 fn default_runtime() -> String {
     "claude".to_string()
-}
-
-/// Role type for an agent.
-#[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum AgentRole {
-    Actor,
-    Reviewer,
-}
-
-impl std::fmt::Display for AgentRole {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Actor => write!(f, "actor"),
-            Self::Reviewer => write!(f, "reviewer"),
-        }
-    }
-}
-
-impl std::str::FromStr for AgentRole {
-    type Err = String;
-    fn from_str(s: &str) -> std::result::Result<Self, Self::Err> {
-        match s {
-            "actor" => Ok(Self::Actor),
-            "reviewer" => Ok(Self::Reviewer),
-            _ => Err(format!(
-                "unknown AgentRole: {s}. Expected 'actor' or 'reviewer'."
-            )),
-        }
-    }
-}
-
-/// A parsed agent definition from a `.md` file.
-#[derive(Debug, Clone, serde::Serialize)]
-pub struct AgentDef {
-    /// Agent identifier (from file stem).
-    pub name: String,
-    /// Role type: actor or reviewer.
-    pub role: AgentRole,
-    /// Whether this agent is permitted to commit code.
-    pub can_commit: bool,
-    /// Optional model override.
-    pub model: Option<String>,
-    /// The runtime to use for this agent (defaults to "claude").
-    pub runtime: String,
-    /// The prompt template (full markdown body after frontmatter).
-    pub prompt: String,
 }
 
 /// Parse an agent `.md` file into an `AgentDef`.
