+++ conductor-core/src/config.rs
@@ -5,85 +5,9 @@ use std::sync::OnceLock;
 
 use crate::error::{ConductorError, Result};
 
-/// Controls which permission flag is passed to Claude Code when launching agent runs.
-///
-/// ```toml
-/// [general]
-/// agent_permission_mode = "skip-permissions" # default — uses --dangerously-skip-permissions
-/// agent_permission_mode = "auto-mode"        # uses --enable-auto-mode (may prompt in headless agents)
-/// ```
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "kebab-case")]
-pub enum AgentPermissionMode {
-    /// Use `--enable-auto-mode` (may prompt for permissions in headless agents).
-    AutoMode,
-    /// Use `--dangerously-skip-permissions` (default for headless agent runs).
-    #[default]
-    SkipPermissions,
-    /// Use `--permission-mode plan` (read-only mode for repo-scoped agents).
-    Plan,
-    /// Use `--dangerously-skip-permissions` + `--allowedTools` read-safe pattern.
-    /// Excludes file-writing tools (Edit, Write, MultiEdit, NotebookEdit) at the
-    /// Claude tool level without locking into plan-mode's "propose before acting"
-    /// flow, so Bash/gh remain fully executable.
-    RepoSafe,
-}
-
-impl AgentPermissionMode {
-    /// Returns the conductor CLI flag for this mode (used in `conductor agent run` passthrough args).
-    pub fn cli_flag(&self) -> &str {
-        match self {
-            Self::AutoMode => "--enable-auto-mode",
-            Self::SkipPermissions => "--dangerously-skip-permissions",
-            Self::Plan => "--permission-mode",
-            Self::RepoSafe => "--permission-mode",
-        }
-    }
-
-    /// Returns the optional value argument that follows the conductor CLI flag.
-    pub fn cli_flag_value(&self) -> Option<&str> {
-        match self {
-            Self::Plan => Some("plan"),
-            Self::RepoSafe => Some("repo-safe"),
-            _ => None,
-        }
-    }
-
-    /// Returns the actual permission flag to pass to the `claude` subprocess.
-    ///
-    /// This differs from `cli_flag()` for `RepoSafe`: conductor receives
-    /// `--permission-mode repo-safe`, but claude receives `--dangerously-skip-permissions`.
-    pub fn claude_permission_flag(&self) -> &str {
-        match self {
-            Self::AutoMode => "--enable-auto-mode",
-            Self::SkipPermissions => "--dangerously-skip-permissions",
-            Self::Plan => "--permission-mode",
-            Self::RepoSafe => "--dangerously-skip-permissions",
-        }
-    }
-
-    /// Returns the optional value argument that follows the claude permission flag.
-    pub fn claude_permission_flag_value(&self) -> Option<&str> {
-        match self {
-            Self::Plan => Some("plan"),
-            _ => None,
-        }
-    }
-
-    /// Returns the `--allowedTools` pattern for this mode, if any.
-    ///
-    /// Plan and RepoSafe modes allow read-only and shell tools (Bash, Glob, Grep, Read,
-    /// WebFetch, WebSearch) plus all MCP tools, while excluding file-writing tools
-    /// (Edit, Write, MultiEdit, NotebookEdit).
-    pub fn allowed_tools(&self) -> Option<&'static str> {
-        match self {
-            Self::Plan | Self::RepoSafe => {
-                Some("Bash,Glob,Grep,Read,WebFetch,WebSearch,mcp__conductor__*,mcp__*")
-            }
-            _ => None,
-        }
-    }
-}
+// Re-export moved types from runkon-runtimes
+pub use runkon_runtimes::config::RuntimeConfig;
+pub use runkon_runtimes::permission::PermissionMode as AgentPermissionMode;
 
 /// Controls whether an agent is auto-started after creating a worktree from a ticket.
 #[derive(Debug, Clone, Copy, Default, PartialEq, Eq, Serialize, Deserialize)]
@@ -223,49 +147,6 @@ pub struct HookConfig {
 pub struct NotifyConfig {
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
     pub hooks: Vec<HookConfig>,
-}
-
-/// Configuration for a named agent runtime (RFC 007).
-///
-/// # Example: Gemini CLI
-/// ```toml
-/// [runtimes.gemini]
-/// type = "cli"
-/// binary = "gemini"
-/// args = ["-m", "{{model}}", "-p", "{{prompt}}", "--approval-mode=yolo"]
-/// default_model = "gemini-2.5-flash"
-/// result_field = "response"
-/// token_fields = "stats.models.*.tokens.total"
-/// ```
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-pub struct RuntimeConfig {
-    /// "cli", "api", or "script". Defaults to "cli".
-    #[serde(rename = "type", default, skip_serializing_if = "Option::is_none")]
-    pub runtime_type: Option<String>,
-    /// For "cli": binary name (e.g. "gemini"). Must be on PATH.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub binary: Option<String>,
-    /// For "cli": arg template. {{prompt}} and {{model}} are substituted.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub args: Option<Vec<String>>,
-    /// For "cli": how to pass the prompt. "arg" (default) or "stdin".
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub prompt_via: Option<String>,
-    /// Default model ID passed as {{model}}. Overridden by agent frontmatter `model:`.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub default_model: Option<String>,
-    /// Dot-path into JSON stdout to extract result_text (e.g. "response").
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub result_field: Option<String>,
-    /// Dot-path into JSON stdout to extract total token count (optional).
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub token_fields: Option<String>,
-    /// For "api": env var name holding the API key.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub api_key_env: Option<String>,
-    /// For "script": shell command to execute.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub command: Option<String>,
 }
 
 #[derive(Debug, Clone, Default, Serialize, Deserialize)]
