+++ conductor-core/src/workflow/claude_agent_executor.rs
@@ -1,7 +1,11 @@
+use std::path::PathBuf;
+use std::sync::Arc;
+
 use crate::agent::AgentRunStatus;
-use crate::config::Config;
+use crate::config::{agent_log_path, Config};
 use crate::error::{ConductorError, Result};
-use crate::runtime::PollError;
+use crate::runtime::adapter::SqliteHostAdapter;
+use crate::runtime::{PollError, RuntimeOptions};
 use crate::workflow::action_executor::{
     ActionExecutor, ActionOutput, ActionParams, ExecutionContext,
 };
@@ -45,7 +49,30 @@ impl ActionExecutor for ClaudeAgentExecutor {
 
         let (agent_def, prompt) = super::helpers::load_agent_and_build_prompt(ectx, params)?;
 
-        let runtime = crate::runtime::resolve_runtime(&agent_def.runtime, &self.config)?;
+        let options = RuntimeOptions {
+            binary_path: std::env::current_exe()
+                .ok()
+                .and_then(|p| {
+                    let sibling = p.parent()?.join("conductor");
+                    sibling.exists().then(|| sibling)
+                })
+                .unwrap_or_else(|| PathBuf::from("conductor")),
+            log_path_for_run: Arc::new(|run_id| {
+                agent_log_path(run_id).unwrap_or_else(|_| std::env::temp_dir().join(format!("{run_id}.log")))
+            }),
+            workspace_root: self.config.general.workspace_root.clone(),
+        };
+
+        let runtime = crate::runtime::resolve_runtime(
+            &agent_def.runtime,
+            self.config.general.agent_permission_mode,
+            &self.config.runtimes,
+            &options,
+        )?;
+
+        let tracker = Arc::new(SqliteHostAdapter::new(ectx.db_path.clone()));
+        let event_sink = tracker.clone();
+        let log_path = agent_log_path(&ectx.run_id)?;
 
         let request = crate::runtime::RuntimeRequest {
             run_id: ectx.run_id.clone(),
@@ -55,7 +82,9 @@ impl ActionExecutor for ClaudeAgentExecutor {
             model: ectx.model.clone(),
             bot_name: ectx.bot_name.clone(),
             plugin_dirs: ectx.plugin_dirs.clone(),
-            db_path: ectx.db_path.clone(),
+            tracker,
+            event_sink,
+            log_path,
         };
 
         runtime.spawn_validated(&request)?;
@@ -64,7 +93,6 @@ impl ActionExecutor for ClaudeAgentExecutor {
             &ectx.run_id,
             ectx.shutdown.as_ref(),
             ectx.step_timeout,
-            &ectx.db_path,
         ) {
             Ok(run) => run,
             Err(PollError::Cancelled) => {
