+++ runkon-runtimes/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "runkon-runtimes"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+description = "Portable agent runtime harness — spawn, poll, and cancel agents without depending on conductor's full domain"
+
+[features]
+rusqlite = ["dep:rusqlite"]
+utoipa = ["dep:utoipa"]
+
+[dependencies]
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+thiserror = "2"
+tracing = "0.1"
+libc = "0.2"
+rusqlite = { version = "0.32", features = ["bundled"], optional = true }
+utoipa = { version = "5", optional = true }
+
+[dev-dependencies]
+tempfile = "3"
