+++ runkon-runtimes/src/lib.rs
@@ -1,0 +1,21 @@
+//! Portable agent runtime harness — spawn, poll, and cancel agents
+//! without depending on conductor's full domain.
+
+pub mod agent_def;
+pub mod config;
+pub mod error;
+pub mod headless;
+pub mod permission;
+pub mod process_utils;
+pub mod run;
+pub mod runtime;
+pub mod text_util;
+pub mod tracker;
+
+pub use agent_def::{AgentDef, AgentRole};
+pub use config::RuntimeConfig;
+pub use error::{Result, RuntimeError};
+pub use permission::PermissionMode;
+pub use run::{AgentRun, AgentRunStatus, PlanStep, StepStatus};
+pub use runtime::{AgentRuntime, PollError, RuntimeOptions, RuntimeRequest};
+pub use tracker::{NoopEventSink, RunEventSink, RunTracker, RuntimeEvent};
